+++ perun/cli.py
@@ -883,7 +883,39 @@ def parse_yaml_single_param(ctx, param, value):
     return unit_to_params
 
 
+def parse_minor_version(ctx, _, value):
+    """Callback function for parsing the minor version list for running the automation
+
+    :param Context ctx: context of the called command
+    :param click.Option _: parameter that is being parsed and read from commandline
+    :param str value: value that is being read from the commandline
+    :returns list: list of MinorVersion objects
+    """
+    minors = []
+    if value:
+        pcs = PCS(store.locate_perun_dir_on(os.getcwd()))
+        for minor_version in value:
+            massaged_version = vcs.massage_parameter(pcs.vcs_type, pcs.vcs_path, minor_version)
+            # If we should crawl all of the parents, we collect them
+            if ctx.params['crawl_parents']:
+                minors.extend(vcs.walk_minor_versions(
+                    pcs.vcs_type, pcs.vcs_path, massaged_version
+                ))
+            # Otherwise we retrieve the minor version info for the param
+            else:
+                minors.append(vcs.get_minor_version_info(
+                    pcs.vcs_type, pcs.vcs_path, massaged_version
+                ))
+    return minors
+
+
 @cli.group()
+@click.option('--minor-version', '-m', 'minor_version_list', nargs=1, multiple=True,
+              callback=parse_minor_version, default=['HEAD'],
+              help='Specifies the head minor version, for which the profiles will be collected.')
+@click.option('--crawl-parents', '-c', is_flag=True, default=False, is_eager=True,
+              help='If set to true, then for each specified minor versions, profiles for parents'
+                   ' will be collected as well')
 @click.option('--cmd', '-c', nargs=1, required=False, multiple=True, default=[''],
               help='Command that is being profiled. Either corresponds to some'
               ' script, binary or command, e.g. ``./mybin`` or ``perun``.')
@@ -965,16 +997,25 @@ def init_unit_commands(lazy_init=True):
                    ' This way the file with collected data will have a resulting filename w.r.t '
                    ' to this parameter. Refer to :ckey:`format.output_profile_template` for more'
                    ' details about the format of the template.')
-def run(**kwargs):
+@click.option('--minor-version', '-m', 'minor_version_list', nargs=1, multiple=True,
+              callback=parse_minor_version, default=['HEAD'],
+              help='Specifies the head minor version, for which the profiles will be collected.')
+@click.option('--crawl-parents', '-c', is_flag=True, default=False, is_eager=True,
+              help='If set to true, then for each specified minor versions, profiles for parents'
+                   ' will be collected as well')
+@click.pass_context
+def run(ctx, **kwargs):
     """Generates batch of profiles w.r.t. specification of list of jobs.
 
     Either runs the job matrix stored in local.yml configuration or lets the
     user construct the job run using the set of parameters.
     """
+    ctx.obj = kwargs
 
 
 @run.command()
-def matrix(**kwargs):
+@click.pass_context
+def matrix(ctx, **kwargs):
     """Runs the jobs matrix specified in the local.yml configuration.
 
     This commands loads the jobs configuration from local configuration, builds
@@ -992,6 +1033,7 @@ def matrix(**kwargs):
     the job matrix inside local configuration and to :doc:`config` how to work
     with Perun's configuration files.
     """
+    kwargs.update({'minor_version_list': ctx.obj['minor_version_list']})
     runner.run_matrix_job(**kwargs)
 
 
@@ -1042,7 +1084,8 @@ def parse_yaml_param(ctx, param, value):
               callback=parse_yaml_param,
               help='Additional parameters for the <postprocessor> read from the'
               ' file in YAML format')
-def job(**kwargs):
+@click.pass_context
+def job(ctx, **kwargs):
     """Run specified batch of perun jobs to generate profiles.
 
     This command correspond to running one isolated batch of profiling jobs,
@@ -1095,6 +1138,7 @@ def job(**kwargs):
     postprocessors refer to :ref:`collectors-list` and
     :ref:`postprocessors-list` respectively.
     """
+    kwargs.update({'minor_version_list': ctx.obj['minor_version_list']})
     runner.run_single_job(**kwargs)